project folders restructured and ahhf logo and background added in assets
