+++ pubspec.yaml
@@ -65,7 +65,7 @@ flutter:
 
 
   assets:
-    - images/
+    - assets/images/
 
 
   # An image asset can refer to one or more resolution-specific "variants", see
